+++ src/models/key/key-model.ts
@@ -2,7 +2,9 @@ import { ModelFactory } from 'meseret'
 
 import { keyPaths } from './key-paths'
 import { keyStatics } from './key-statics'
+import { getMeseretUtilsPreState } from '../../pre-state'
 
+// todo: enumerate...
 export type IKeyPurpose = 'PASSWORD_RESET'
 export const keyPurposes: IKeyPurpose[] = ['PASSWORD_RESET']
 
@@ -16,6 +18,7 @@ export interface IKey {
 }
 
 export const keyModelFactory = new ModelFactory<IKey, {}, typeof keyStatics>({
+  mongooseInstance: getMeseretUtilsPreState().mongooseInstance,
   name: 'key',
   paths: keyPaths,
   statics: keyStatics